- Fix home tab and payment (decimals)

+++ app/src/main/java/com/kibou/abisoyeoke_lawal/coupinapp/models/GeneralModels.kt
@@ -156,8 +156,8 @@ data class Multiple (
 )
 
 data class Price (
-    val old: Long?,
-    val new: Long?
+    val old: Double?,
+    val new: Double?
 )
 
 data class Review (
